fix: consolidate formatters
