+++ src/formatters/deleteResultFormatter.ts
@@ -4,30 +4,17 @@
  * Licensed under the BSD 3-Clause license.
  * For full license text, see LICENSE.txt file in the repo root or https://opensource.org/licenses/BSD-3-Clause
  */
-import * as os from 'os';
 import { ux } from '@oclif/core';
 import * as chalk from 'chalk';
-import { dim, underline } from 'chalk';
-import { CodeCoverageWarnings, DeployResult, FileResponse, RequestStatus } from '@salesforce/source-deploy-retrieve';
+import { DeployResult, FileResponse, RequestStatus } from '@salesforce/source-deploy-retrieve';
 import { ensureArray } from '@salesforce/kit';
 import { bold } from 'chalk';
 import { StandardColors } from '@salesforce/sf-plugins-core';
-import { DeleteSourceJson, Formatter, TestLevel, Verbosity } from '../utils/types';
-import {
-  sortFileResponses,
-  asRelativePaths,
-  tableHeader,
-  sortTestResults,
-  error,
-  success,
-  check,
-} from '../utils/output';
-import { coverageOutput } from '../utils/coverage';
+import { DeleteSourceJson, Formatter, TestLevel } from '../utils/types';
+import { sortFileResponses, asRelativePaths } from '../utils/output';
+import { TestResultsFormatter } from '../formatters/testResultsFormatter';
 
-export class DeleteResultFormatter implements Formatter<DeleteSourceJson> {
-  private testLevel: TestLevel | undefined;
-  private verbosity: Verbosity;
-
+export class DeleteResultFormatter extends TestResultsFormatter implements Formatter<DeleteSourceJson> {
   public constructor(
     protected result: DeployResult,
     protected flags: Partial<{
@@ -35,6 +22,7 @@ export class DeleteResultFormatter implements Formatter<DeleteSourceJson> {
       verbose: boolean;
     }>
   ) {
+    super(result, flags);
     this.testLevel = flags['test-level'];
     this.verbosity = this.determineVerbosity();
   }
@@ -104,85 +92,6 @@ export class DeleteResultFormatter implements Formatter<DeleteSourceJson> {
     } else {
       this.displayFailures();
     }
-  }
-
-  private displayTestResults(): void {
-    if (this.testLevel === TestLevel.NoTestRun || !this.result.response.runTestsEnabled) {
-      ux.log();
-      return;
-    }
-
-    this.displayVerboseTestFailures();
-
-    if (this.verbosity === 'verbose') {
-      this.displayVerboseTestSuccesses();
-      this.displayVerboseTestCoverage();
-    }
-
-    ux.log();
-    ux.log(tableHeader('Test Results Summary'));
-    ux.log(`Passing: ${this.result.response.numberTestsCompleted ?? 0}`);
-    ux.log(`Failing: ${this.result.response.numberTestErrors ?? 0}`);
-    ux.log(`Total: ${this.result.response.numberTestsTotal ?? 0}`);
-    const time = this.result.response.details.runTestResult?.totalTime ?? 0;
-    if (time) ux.log(`Time: ${time}`);
-    // I think the type might be wrong in SDR
-    ensureArray(this.result.response.details.runTestResult?.codeCoverageWarnings).map(
-      (warning: CodeCoverageWarnings & { name?: string }) =>
-        ux.warn(`${warning.name ? `${warning.name} - ` : ''}${warning.message}`)
-    );
-  }
-
-  private displayVerboseTestCoverage(): void {
-    const codeCoverage = ensureArray(this.result.response.details.runTestResult?.codeCoverage);
-    if (codeCoverage.length) {
-      const coverage = codeCoverage.sort((a, b) => (a.name.toUpperCase() > b.name.toUpperCase() ? 1 : -1));
-      ux.log();
-      ux.log(tableHeader('Apex Code Coverage'));
-
-      ux.table(coverage.map(coverageOutput), {
-        name: { header: 'Name' },
-        numLocations: { header: '% Covered' },
-        lineNotCovered: { header: 'Uncovered Lines' },
-      });
-    }
-  }
-
-  private displayVerboseTestSuccesses(): void {
-    const successes = ensureArray(this.result.response.details.runTestResult?.successes);
-    if (successes.length > 0) {
-      const testSuccesses = sortTestResults(successes);
-      ux.log();
-      ux.log(success(`Test Success [${successes.length}]`));
-      for (const test of testSuccesses) {
-        const testName = underline(`${test.name}.${test.methodName}`);
-        ux.log(`${check} ${testName}`);
-      }
-    }
-  }
-
-  private displayVerboseTestFailures(): void {
-    if (!this.result.response.numberTestErrors) return;
-    const failures = ensureArray(this.result.response.details.runTestResult?.failures);
-    const failureCount = this.result.response.details.runTestResult?.numFailures;
-    const testFailures = sortTestResults(failures);
-    ux.log();
-    ux.log(error(`Test Failures [${failureCount}]`));
-    for (const test of testFailures) {
-      const testName = underline(`${test.name}.${test.methodName}`);
-      ux.log(`• ${testName}`);
-      ux.log(`  ${dim('message')}: ${test.message}`);
-      if (test.stackTrace) {
-        const stackTrace = test.stackTrace.replace(/\n/g, `${os.EOL}    `);
-        ux.log(`  ${dim('stacktrace')}: ${os.EOL}    ${stackTrace}`);
-      }
-      ux.log();
-    }
-  }
-
-  private determineVerbosity(): Verbosity {
-    if (this.flags.verbose) return 'verbose';
-    return 'normal';
   }
 
   private displayFailures(): void {
